% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{delete_artefact}
\alias{delete_artefact}
\title{Delete an artefact (or a whole project)}
\usage{
delete_artefact(project, artefact = "all")
}
\arguments{
\item{project}{Project name.}

\item{artefact}{Either a canonical artefact name, or \code{"all"} to
remove the entire project directory.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Delete an artefact (or a whole project)
}
