% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{list_projects}
\alias{list_projects}
\title{List existing project names}
\usage{
list_projects()
}
\value{
Character vector of project names (immediate subdirectories of
\verb{<data_root()>/projects/}).
}
\description{
List existing project names
}
