% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{load_artefact}
\alias{load_artefact}
\title{Load an artefact from a project directory}
\usage{
load_artefact(project, artefact, default = NULL)
}
\arguments{
\item{project}{Project name.}

\item{artefact}{One of the canonical artefact names.}

\item{default}{Value to return if the artefact does not exist.}
}
\value{
The stored object, or \code{default}.
}
\description{
Load an artefact from a project directory
}
