% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_pilot.R
\name{pilot_job_cancel}
\alias{pilot_job_cancel}
\title{Cancel a running pilot job}
\usage{
pilot_job_cancel(handle)
}
\arguments{
\item{handle}{Callr process handle (from \code{start_pilot_job()$handle}).}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Cancel a running pilot job
}
