% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_cache_dir}
\alias{project_cache_dir}
\title{Return the cache directory for a project (used by \code{rank_records})}
\usage{
project_cache_dir(project)
}
\arguments{
\item{project}{Project name.}
}
\value{
Character path.
}
\description{
Return the cache directory for a project (used by \code{rank_records})
}
\keyword{internal}
