% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_pull.R
\name{pull_job_cancel}
\alias{pull_job_cancel}
\title{Cancel a running pull job}
\usage{
pull_job_cancel(handle)
}
\arguments{
\item{handle}{Callr process object (from \code{start_pull_job()$handle}).}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Cancel a running pull job
}
