% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rank_job_cancel}
\alias{rank_job_cancel}
\title{Cancel a running ranking job}
\usage{
rank_job_cancel(handle)
}
\arguments{
\item{handle}{Callr process object.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Requires the caller to have kept the \code{handle} returned by
\code{start_rank_job()}.
}
\keyword{internal}
